# Backend/backend/src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: Beschreibung der Rest-Schnitstelle der Healthcare API
  version: 1.0.0
  title: Healthcare App
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /v1
...
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /roles:
    get:
      tags:
        - role
      summary: Alle rollen
      description: Alle rollen
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Role'
    post:
      tags:
        - role
      summary: Erstelle neue Role
      description: Neue Role
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Role'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Role'
  '/roles/{roleId}':
    parameters:
      - name: roleId
        in: path
        required: true
        type: string
        format: string
    get:
      tags:
        - role
      summary: Einzelne role
      description: Einzelne Role
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Role'
    put:
      tags:
        - role
      summary: Role bearbeiten
      description: Role Bearbeiten
      responses:
        '200':
          description: OK
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Role'
    delete:
        tags:
          - Role
        summary: löschen
        description: löschen
        responses:
          '200':
            description: OK
          '402':
            description: NOT_FOUND
  /users:
    post:
      tags:
        - user
      summary: Erstelle einen neuen user
      description: Einen Neuen User erstellen
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
    get:
      tags:
        - user
      summary: Liste aller User
      description: Alle User
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
  '/users/{userId}':
    parameters:
      - name: userId
        in: path
        required: true
        type: string
        format: string
    get:
      tags:
        - user
      summary: Erhalte informationen über einen User
      description: Ein User
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
    put:
      tags:
        - user
      summary: Einen User bearbeiten
      description: Einen User bearbeiten
      responses:
        '200':
          description: OK
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
    delete:
        tags:
          - User
        summary: löschen
        description: löschen
        responses:
          '200':
            description: OK
          '402':
            description: NOT_FOUND
...
definitions:
...
  User:
    type: object
    required:
      - name
      - password
      - email
      - lastUpdate
      - role
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        format: string
      password:
        type: string
        format: password
      email:
        type: string
        format: string
      lastUpdate:
        type: string
        format: dateTime
      role:
        $ref: '#/definitions/Role'
  Role:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        format: string
...